refactor: rename component and move props
rename Toastr to QToastr, ToastrContainer to Toastr, move those props related to view to from Toastr
to QToastr
