+++ packages/toastr/src/react-component-node-modules/components/Toastr/Toastr.md
@@ -1,7 +1,7 @@
 
 ```js
 <div>
-  <Toastr toastrStyle="qts">{['a', 'b']}</Toastr>
+  <Toastr toastrStyle="qts">已複製到剪貼簿。</Toastr>
   <Toastr toastrStyle="photo-st">PhotoStation</Toastr>
   <Toastr toastrStyle="music-st">MusicStation</Toastr>
   <Toastr toastrStyle="file-st">FileStation</Toastr>
