+++ packages/toastr/src/react-component-node-modules/components/Toastr/package.json
@@ -1,5 +1,5 @@
 {
   "private": true,
   "name": "@react-qui/toastr",
-  "main": "./Toastr.js"
+  "main": "./QToastr.js"
 }
