+++ packages/toastr/src/react-component-node-modules/components/Toastr/styles/Toastr.scss
@@ -1,41 +1,34 @@
-@import 'colors';
-@import 'mixins';
-
-.toastr {
+.top-right {
   display: flex;
-  flex-wrap: nowrap;
-  align-items: center;
+  align-items: flex-start;
+  justify-content: flex-end;
+  margin-top: 24px;
+  margin-right: 24px;
+  pointer-events: none;
+}
+
+.top {
+  display: flex;
+  align-items: flex-start;
   justify-content: center;
-  min-width: 64px;
-  padding: 24px;
-  margin-top: 8px;
-  font-family: Roboto, Helvetica, Arial, Sans-serif, Verdana, Tahoma;
-  font-size: 16px;
-  line-height: 20px;
-  color: $toastr-font-color;
-  text-align: center;
+  margin-top: 24px;
+  pointer-events: none;
 }
 
-.qts {
-  background-color: $qts-default-color;
+.top-left {
+  display: flex;
+  align-items: flex-start;
+  justify-content: flex-start;
+  margin-top: 24px;
+  margin-left: 24px;
+  pointer-events: none;
 }
 
-.photo-st {
-  background-color: $photo-station-color;
-}
-
-.music-st {
-  background-color: $music-station-color;
-}
-
-.file-st {
-  background-color: $file-station-color;
-}
-
-.download-st {
-  background-color: $download-station-color;
-}
-
-.video-st {
-  background-color: $video-station-color;
+.bottom {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  justify-content: flex-end;
+  margin-top: 100px;
+  pointer-events: none;
 }
