--- packages/toastr/src/react-component-node-modules/components/Toastr/styles/ToastrContainer.scss
@@ -1,25 +1,0 @@
-.top-right {
-  display: flex;
-  align-items: flex-start;
-  justify-content: flex-end;
-  margin-top: 24px;
-  margin-right: 24px;
-  pointer-events: none;
-}
-
-.top {
-  display: flex;
-  align-items: flex-start;
-  justify-content: center;
-  margin-top: 24px;
-  pointer-events: none;
-}
-
-.top-left {
-  display: flex;
-  align-items: flex-start;
-  justify-content: flex-start;
-  margin-top: 24px;
-  margin-left: 24px;
-  pointer-events: none;
-}
